% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{basis_rcpp}
\alias{basis_rcpp}
\title{Create a basis vector}
\usage{
basis_rcpp(matvec, resvec, onesvec, theta)
}
\arguments{
\item{matvec}{A single integer.}

\item{resvec}{A single integer.}

\item{onesvec}{A single integer.}

\item{theta}{A single integer.}
}
\description{
Create a basis vector
}
