% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLCE.R
\name{DML}
\alias{DML}
\title{Double Machine Learning
Run Double Machine Learning Method with continuous treatment}
\usage{
DML(
  Y,
  D,
  X,
  n_split = 2,
  method_out = "RF",
  method_treat = "RF",
  fit.lm = "IV",
  ...
)
}
\arguments{
\item{Y}{Outcome vector.}

\item{D}{Treatment variable (scalar).}

\item{X}{Covariates.}

\item{method_out}{Method used to predict the outcome. Random forest (\code{"RF"}) and LASSO (\code{"LASSO"}) are allowed.}

\item{method_treat}{See \code{method_out} options.}
}
\description{
Double Machine Learning
Run Double Machine Learning Method with continuous treatment
}
\examples{
     rfrf_ce <- DML(Y = obs, D = treat, X = X, method_out = 'RF', method_treat = 'RF')
     rfls_ce <- DML(Y = obs, D = treat, X = X, method_out = 'RF', method_treat = 'LASSO')
     lsrf_ce <- DML(Y = obs, D = treat, X = X, method_out = 'LASSO', method_treat = 'RF')
     lsls_ce <- DML(Y = obs, D = treat, X = X, method_out = 'LASSO', method_treat = 'LASSO')
}
