% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastres_cpp}
\alias{fastres_cpp}
\title{Fast partialing out using weighted least squares}
\usage{
fastres_cpp(X, y, w)
}
\arguments{
\item{X}{A single integer.}

\item{y}{A single integer.}

\item{w}{A single integer.}
}
\description{
Fast partialing out using weighted least squares
}
