% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLCE.R
\name{plce}
\alias{plce}
\title{Sparse regression for experimental and observational data.}
\usage{
plce(y, treat, X, inner.boot = 30, num.fit = 25, var.type = "HC3")
}
\arguments{
\item{y}{Vector of the dependent, or outcome variable.}

\item{treat}{Treatment variable, for which the average partial effect is desired. 
May be continous, binary, or count.}

\item{X}{Matrix of covariates. Typical vocabulary would refer to these as
"pre-treatment" covariates.}

\item{inner.boot}{The number of bootstrapped estimates to do in each subsample.  Default is 30.}

\item{num.fit}{The number of cross-fitting iterations to run.  Default is 25.
inner.boot = 30,}

\item{var.type}{The type of variance estimate, passed to \code{vcovHC} in sandwich. Options are
\code{HC0}, \code{HC1}, \code{HC2}, \code{HC3}, with \code{HC3} the default.}
}
\value{
\describe{ 
\item{point}{The estimated average partial effect.}
\item{se}{Standard error of the estimate of the average partial effect.}

\item{sens}{Results from the sensitivity analysis.}
}
}
\description{
Sparse regression for experimental and observational data.
}
\examples{
\dontrun{

}

}
\references{
Ratkovic, Marc.  2020. "Relaxing Assumptions, Improving Inference: 
Utilizing Machine Learning for Valid Causal Inference."  Working Paper.
}
