% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsandsummaries.R
\name{pos_plot}
\alias{pos_plot}
\title{Assessing Positivity of a Treatment Variable}
\usage{
pos_plot(obj, trim = 0.025)
}
\arguments{
\item{obj}{A PLCE object.}

\item{trim}{The percentage to trim off each end of the excess kurtosis statistic.
This helps in reducing}
}
\value{
\describe{
\item{diff}{The estimated excess kurtosis per datum.}
}
}
\description{
Assessing Positivity of a Treatment Variable
}
\references{
Ratkovic, Marc.  2021. "Relaxing Assumptions, Improving Inference:
Utilizing Machine Learning for Valid Causal Inference."  Working Paper.
}
