% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsEMUpdateFunctions.R
\name{qfunction}
\alias{qfunction}
\title{Q Function}
\usage{
qfunction(
  y,
  X,
  beta,
  sigma.sq,
  lambda,
  Ewtsqtausq,
  Etausqinv,
  re.obj,
  prior_covariance
)
}
\description{
Q Function
}
