% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsEMUpdateFunctions.R
\name{updatebeta}
\alias{updatebeta}
\title{update beta}
\usage{
updatebeta(
  y,
  Dtau,
  sigma.sq,
  EM,
  XprimeX,
  Xprimey,
  eig1,
  type,
  blockupdate,
  beta,
  trim,
  i.iter,
  burnin,
  beta.mean.last,
  re.obj,
  tauthresh,
  betathresh
)
}
\description{
update beta
}
