% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsEMUpdateFunctions.R
\name{updateprec}
\alias{updateprec}
\title{update sigma.sq}
\usage{
updateprec(
  y,
  fits,
  beta,
  Dtau,
  EM,
  type,
  rate.offset,
  i.iter,
  Etausqinv,
  re.obj
)
}
\description{
update sigma.sq
}
