% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsEMUpdateFunctions.R
\name{updatetau.gibbs}
\alias{updatetau.gibbs}
\title{update tausq}
\usage{
updatetau.gibbs(
  lambda,
  lambda.sq,
  wts,
  sigma.sq,
  beta,
  tau.last,
  i.iter,
  burnin
)
}
\description{
update tausq
}
