% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsEMUpdateFunctions.R
\name{updatewts.EM}
\alias{updatewts.EM}
\title{update weights via EM}
\usage{
updatewts.EM(x, seq.obj0 = seq.obj)
}
\arguments{
\item{x}{a vector with m in the first element and gamma0 in the second}
}
\description{
update weights via EM
}
\details{
A lookup table with pre-computed integrals, `precomp_wts`, is used internally.
The table is generated in the data-raw directory.
}
